Readapt OZ to 0.1.0; refactor
