+++ src/account-utils.ts
@@ -1,6 +1,5 @@
-import { Numeric, StarknetContract, StringMap } from "./types";
-import { Call, hash, RawCalldata } from "starknet";
-import { BigNumberish, toBN } from "starknet/utils/number";
+import { StarknetContract, StringMap } from "./types";
+import { toBN } from "starknet/utils/number";
 import * as ellipticCurve from "starknet/utils/ellipticCurve";
 import { ec } from "elliptic";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
@@ -9,8 +8,7 @@ import path from "path";
 import {
     ABI_SUFFIX,
     ACCOUNT_CONTRACT_ARTIFACTS_ROOT_PATH,
-    GITHUB_ACCOUNT_ARTIFACTS_URL,
-    TransactionHashPrefix
+    GITHUB_ACCOUNT_ARTIFACTS_URL
 } from "./constants";
 import axios from "axios";
 import { flattenStringMap } from "./utils";
@@ -19,13 +17,6 @@ export type CallParameters = {
     toContract: StarknetContract;
     functionName: string;
     calldata?: StringMap;
-};
-
-type executeCallParameters = {
-    to: bigint;
-    selector: BigNumberish;
-    data_offset: number;
-    data_len: number;
 };
 
 type KeysType = {
@@ -56,86 +47,6 @@ export function signMultiCall(
         return [BigInt(0), BigInt(0)];
     }
     return ellipticCurve.sign(keyPair, BigInt(messageHash).toString(16)).map(BigInt);
-}
-
-/**
- * Prepares the calldata and hashes the message for the multicall execution
- *
- * @param accountAddress address of the account contract
- * @param callParameters array witht the call parameters
- * @param nonce current nonce
- * @param maxFee the maximum fee amoutn set for the contract interaction
- * @param version the transaction version
- * @param chainId the chain identifier
- * @param executionFunctionName the name of the cairo function that performs the execution
- * @returns the message hash for the multicall and the arguments to execute it with
- */
-export function handleMultiInteract(
-    accountAddress: string,
-    callParameters: CallParameters[],
-    nonce: Numeric,
-    maxFee: Numeric,
-    version: Numeric,
-    chainId: Numeric,
-    executionFunctionName: string
-) {
-    // Transform a CallParameters array into Call array, so it can be used by the hash functions
-    const callArray: Call[] = callParameters.map((callParameters) => {
-        return {
-            contractAddress: callParameters.toContract.address,
-            entrypoint: callParameters.functionName,
-            calldata: callParameters.toContract.adaptInput(
-                callParameters.functionName,
-                callParameters.calldata
-            )
-        };
-    });
-
-    const hashable: Array<BigNumberish> = [callArray.length];
-    const executeCallArray: executeCallParameters[] = [];
-    const rawCalldata: RawCalldata = [];
-
-    // Parse the Call array to create the objects which will be accepted by the contract
-    callArray.forEach((call) => {
-        executeCallArray.push({
-            to: BigInt(call.contractAddress),
-            selector: hash.starknetKeccak(call.entrypoint),
-            data_offset: rawCalldata.length,
-            data_len: call.calldata.length
-        });
-        hashable.push(
-            call.contractAddress,
-            hash.starknetKeccak(call.entrypoint),
-            rawCalldata.length,
-            call.calldata.length
-        );
-        rawCalldata.push(...call.calldata);
-    });
-
-    const adaptedNonce = nonce.toString();
-    const adaptedMaxFee = "0x" + maxFee.toString(16);
-    const adaptedVersion = "0x" + version.toString(16);
-    const adaptedChainId = chainId.toString();
-
-    hashable.push(rawCalldata.length, ...rawCalldata, adaptedNonce);
-    const calldataHash = hash.computeHashOnElements(hashable);
-    const messageHash = hash.computeHashOnElements([
-        TransactionHashPrefix.INVOKE,
-        adaptedVersion,
-        accountAddress,
-        hash.getSelectorFromName(executionFunctionName),
-        calldataHash,
-        adaptedMaxFee,
-        adaptedChainId
-    ]);
-
-    const args = {
-        call_array: executeCallArray,
-        calldata: rawCalldata,
-        nonce: adaptedNonce
-    };
-
-    return { messageHash, args };
 }
 
 export async function handleAccountContractArtifacts(
